feat: separated SearchBar component and configured React Router
