+++ class-components/src/components/Results/Results.tsx
@@ -1,69 +1,24 @@
-import './Results.css';
 import { Episode } from '../../services/DataFetch';
+import './Results.css';
 
 interface ResultsProps {
   episodes: Episode[];
   pageNumber: number;
   totalPages: number;
   onPageChange: (pageNumber: number) => void;
+  onEpisodeClick: (id: string) => void;
 }
 
-/*export class Results extends Component<ResultsProps> {
-  renderEpisodes = (): ReactNode => {
-    const { episodes } = this.props;
-
-    return episodes.map((episode, index) => (
-      <div key={index} className="results__item">
-        <h3 className="results__item-title">{episode.title}</h3>
-        <div className="results__item-info">
-          <div className="season">
-            <b>Season:</b> {episode.seasonNumber},
-          </div>
-          <div className="episode">
-            <b>Episode:</b> {episode.episodeNumber},
-          </div>
-          <div className="series">
-            <b>Series:</b> {episode.seriesTitle}
-          </div>
-        </div>
-      </div>
-    ));
-  };
-
-  renderPagination = (): ReactNode => {
-    const { pageNumber, totalPages, onPageChange } = this.props;
-
-    return (
-      <div className="pagination">
-        <button disabled={pageNumber === 1} onClick={() => onPageChange(pageNumber - 1)}>
-          Previous
-        </button>
-        <span>
-          {pageNumber} / {totalPages}
-        </span>
-        <button disabled={pageNumber === totalPages} onClick={() => onPageChange(pageNumber + 1)}>
-          Next
-        </button>
-      </div>
-    );
-  };
-
-  render(): ReactNode {
-    return (
-      <div className="results">
-        {this.renderEpisodes()}
-        {this.renderPagination()}
-      </div>
-    );
-  }
-}*/
-
-export function Results(props: ResultsProps) {
+export function Results({
+  episodes,
+  pageNumber,
+  totalPages,
+  onPageChange,
+  onEpisodeClick,
+}: ResultsProps) {
   const renderEpisodes = () => {
-    const { episodes } = props;
-
-    return episodes.map((episode, index) => (
-      <div key={index} className="results__item">
+    return episodes.map((episode) => (
+      <div key={episode.uid} className="results__item" onClick={() => onEpisodeClick(episode.uid)}>
         <h3 className="results__item-title">{episode.title}</h3>
         <div className="results__item-info">
           <div className="season">
@@ -81,8 +36,6 @@ export function Results(props: ResultsProps) {
   };
 
   const renderPagination = () => {
-    const { pageNumber, totalPages, onPageChange } = props;
-
     return (
       <div className="pagination">
         <button disabled={pageNumber === 1} onClick={() => onPageChange(pageNumber - 1)}>
