+++ class-components/src/components/SearchBar/SearchBar.tsx
@@ -1,4 +1,4 @@
-import { ChangeEvent, useState } from 'react';
+import React, { ChangeEvent } from 'react';
 import './SearchBar.css';
 import { Input } from '../../utils/ui/Input/Input';
 import { Button } from '../../utils/ui/Button/Button';
@@ -11,24 +11,25 @@ interface SearchBarProps {
   onSearchChange: (searchItem: string) => void;
 }
 
-export function SearchBar(props: SearchBarProps) {
-  const [searchItem, setSearchItem] = useState(props.searchItem);
-  const [warning, setWarning] = useState('');
-
+export const SearchBar: React.FC<SearchBarProps> = ({
+  searchItem,
+  error,
+  setError,
+  onSearch,
+  onSearchChange,
+}) => {
   const handleChange = (event: ChangeEvent<HTMLInputElement>) => {
     const searchItem = event.target.value;
-    setSearchItem(searchItem);
-    setWarning('');
-    props.onSearchChange(searchItem);
+    onSearchChange(searchItem);
   };
 
   const handleSearch = () => {
     const trimmedSearchItem = searchItem.trim();
     const regex = /^[a-zA-Z\s]+$/;
     if (!regex.test(trimmedSearchItem)) {
-      setWarning('Please use only Latin letters.');
+      setError(new Error('Please use only Latin letters.'));
     } else {
-      props.onSearch(trimmedSearchItem, 1);
+      onSearch(trimmedSearchItem, 1);
     }
   };
 
@@ -39,8 +40,7 @@ export function SearchBar(props: SearchBarProps) {
       <Button className="search-bar__search-button" onClick={handleSearch}>
         Search
       </Button>
-      {warning && <div className="warning">{warning}</div>}
-      {props.error && <div className="error">Error: {props.error.message}</div>}
+      {error && <div className="error">Error: {error.message}</div>}
     </div>
   );
-}
+};
