+++ class-components/src/utils/ui/Input/Input.css
@@ -1,4 +1,4 @@
-.search-bar__input {
+input[type="text"] {
   padding: 5px;
   font-size: 24px;
   flex: 1;
